% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cms.R
\name{cms.integration}
\alias{cms.integration}
\title{cms.integration}
\usage{
cms.integration(sce, k, group, embedding = "PCA",
  assay_name = "logcounts", k_min = NA, scale = TRUE, n_pc = 30,
  cell_min = 10)
}
\arguments{
\item{sce}{SingleCellExperiment \code{\link{SingleCellExperiment}} object containing all cells of interest.}

\item{k}{Numeric, defining the number of k-nearest neighbours to use for testing the "mixing" within the neighbourhood.}

\item{group}{Character string specifying the name of variable used to define groups (batches). Should be have a corresonding element in knn.}

\item{embedding}{Character string specifying the embeddings to use as subspace to calculate distance distributions.
The default PCA performs a principle component analysis and uses the n_pc first PCs to define the subspace.}

\item{assay_name}{Character string defining the count assay to use. If not otherwise specified by 'embeddings', a PCA on this assay is performed to calculate a new PCA subspace. Must be one of 'names(assays(sce))'.
Default is 'logcounts' and should not be changed, if other embeddings than PCA should be used.}

\item{k_min}{Numeric giving the minimum number of k nearest neighbours to use. cells to include.
If k_min is defined the overall distance density distribution of knn will be determined and only cells up to the first local minima that includes more than kmin cells will be used.}

\item{scale}{A logical value indicating if cms results should be scaled within each neighbourhood.}

\item{n_pc}{Numeric giving the number of PCs to include to define the subspace to calculate cell distances in.}

\item{cell_min}{Minimum number of cells from each group to be included into the AD test for calculation of cms.
Should be > 10 to make the ad.test function working.}
}
\description{
Calculates cell-specific mixing scores based on euclidean distances within subspace of integrated data.
}
\details{
The cms tests the hypothesis, that group-specific distance distributions of the k-nearest neighbouring cells have the same underlying unspecified distribution using the Anderson-Darling test as implemented in \code{\link{cms.cell}}.
In default the function uses all distances and group label defined in knn. If k_min is specified, the overall distance distribution is checked for the first local minimum, that includes at least kmin cells.
This can be used to adapt to the local structure of the datatset e.g. prevent cells from a distinct different cluster to be included, while keeping the overall number of k nearest neighbours reasonable large.
cms.pca uses a SingleCellExperiment \code{\link{SingleCellExperiment}} object as input and calculates euclidean distances within the subspace defined by the n_pc argument (default = 30).
Generalized version of \code{\link{cms.pca}}. In default both functions are identical.
}
\examples{

}
\seealso{
\code{\link{cms.cell}}, \code{\link{scale.cms}}, \code{\link{cms.pca}}.

Other cms functions: \code{\link{cms.cell}},
  \code{\link{cms.pca}}, \code{\link{cms.summary}},
  \code{\link{filter.locmin}}, \code{\link{scale.cms}}
}
\concept{cms functions}
