% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cms.R
\name{cms.summary}
\alias{cms.summary}
\title{cms.summary}
\usage{
cms.summary(cms_res, sum_var = NULL, sce = NULL)
}
\arguments{
\item{cms_res}{data frame of cms scores to be summarized. Columns should correspond to cms scores, rows to cells.}

\item{sum_var}{variable to group summaries on. Need to correspond to a colData variable}

\item{sce}{SingleCellExperiment object describing the cells specified in cms_res.}
}
\description{
Summarizes cms scores, if specified based on groups
}
\details{
Returns the mean of cms score to make different conditions/methods/.. comparable.
In default the mean of each cloumn of cms_res is returned. Groups to summarize can be specified by sum_var and sce.
}
\examples{

}
\seealso{
\code{\link{compare.integration}}, \code{\link{compare.cluster}}

Other cms functions: \code{\link{cms.cell}},
  \code{\link{cms.integration}}, \code{\link{cms.pca}},
  \code{\link{filter.locmin}}, \code{\link{smooth.cms}}
}
\concept{cms functions}
