% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{compare.cluster}
\alias{compare.cluster}
\title{compare.cluster}
\usage{
compare.cluster(cms_res, cluster_var, cms_var = "cms", sce = NULL,
  violin = TRUE)
}
\arguments{
\item{cms_res}{data frame of cms scores to be summarized. The cms score should be in the first coulmn. It should either contain a second column with corresponding levels for cluster_var or  sce need to be specified.}

\item{cluster_var}{character string specifying the name of the factor level variable to summarize cms scores on.}

\item{cms_var}{Character string specifying the name of the cms_res to use. Default is "cms".}

\item{sce}{SingleCellexperiment object should only be specified if cluster_var is not already provided in cms_res. Should contain 'cluster_var' within colData.
If sce is specified, rownames of cms_res need to correspond to colnames of sce (should contain the same cells).}

\item{violin}{A logical, if true (default) violin plots are provided. Otherwise boxplots are generated.}
}
\description{
Creates summary boxplots of cms scores for different groups/cluster.
}
\details{
Plots summarized cms scores for a specified list. This function is intended to visualize and compare cms scores among clusters or other dataset variables.
}
\examples{

}
\seealso{
\code{\link{compare.integration}}

Other visualize cms functions: \code{\link{compare.integration}},
  \code{\link{vis.cms}}, \code{\link{vis.group}},
  \code{\link{vis.hist}}, \code{\link{vis.overview}}
}
\concept{visualize cms functions}
