% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmsSummary.R
\name{compareCluster}
\alias{compareCluster}
\title{compareCluster}
\usage{
compareCluster(cms_res, cluster_var, cms_var = "cms", sce = NULL,
  violin = FALSE)
}
\arguments{
\item{cms_res}{data frame of cms scores to be summarized. Cms scores should be in the first column.
Either a second column with corresponding levels of \code{cluster_var} or \code{sce} with levels in \code{colData(sce)} need to be specified.}

\item{cluster_var}{Character. Name of the factor level variable to summarize cms scores on.}

\item{cms_var}{Character. Name of the cms_res to use. Default is "cms".}

\item{sce}{\code{SingleCellexperiment} object. Should only be specified if \code{cluster_var} is not already provided in \code{cms_res}.}

\item{violin}{A logical. If true violin plots are plotted, while the default (FALSE) will plot ridge plots.}
}
\value{
a \code{ggplot} object.
}
\description{
Creates summary plots of cms scores for different groups/cluster.
}
\details{
Plots summarized cms scores for a specified list.
This function is intended to visualize and compare cms scores among clusters or other dataset variables.
}
\examples{
library(SingleCellExperiment)

load(system.file("extdata/sim30.rda", package = "CellMixS"))
load(system.file("extdata/cms_sim30.rda", package = "CellMixS"))

sce <- sim_30[[1]][, c(1:50,500:550)]
colData(sce)$group <- sample(c("A", "B", "C"), ncol(sce), replace = TRUE)
compareCluster(cms_sim30, "group", cms_var = "cms", sce = sce, violin = TRUE)

}
\seealso{
\code{\link{compareIntegration}}

Other visualize cms functions: \code{\link{compareIntegration}},
  \code{\link{visCms}}, \code{\link{visGroup}},
  \code{\link{visHist}}, \code{\link{visOverview}}
}
\concept{visualize cms functions}
