% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmsHelper.R
\name{.filterKnn}
\alias{.filterKnn}
\title{.filterKnn}
\usage{
.filterKnn(knn_cell, batch_min, group, sce)
}
\arguments{
\item{knn_cell}{Data frame with one column "distance" and one column named
by the group variable. Rows correspond to the knn cells and do not need
rownames.}

\item{batch_min}{Numeric. Minimum number of cells per batch to include.}

\item{group}{Character. Name of group/batch variable.
Needs to be one of \code{names(knn)}.}

\item{sce}{A \code{SingleCellExperiment} object with the combined data.}
}
\value{
data.frame with two columns (index, distance) for filtered knn cells.
}
\description{
.filterKnn
}
\seealso{
\code{\link{.cmsCell}}

Other helper functions: 
\code{\link{.cmsCell}()},
\code{\link{.defineSubspace}()},
\code{\link{.filterLocMin}()},
\code{\link{.ldfKnn}()},
\code{\link{.smoothCms}()}
}
\concept{helper functions}
