% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cms.R
\name{filter.locmin}
\alias{filter.locmin}
\title{filter.locmin}
\usage{
filter.locmin(knn_cell, k_min)
}
\arguments{
\item{knn_cell}{Data frame with distances and group information of cells. Should have one column named "distance" with distances towards the k nearest neighbours and one column named after the group variable.
Rows correspond to the knn cells and do not need rownames.}

\item{k_min}{Numeric giving the minimum number of k nearest neighbours to use. cells to include.
If k_min is defined the overall distance density distribution of knn will be determined and only cells up to the first local minima that includes more than kmin cells will be used.}
}
\value{

}
\description{
Function to filter number of knn to use for cms calculation by overall distance density distribution.
}
\details{
Internal function to filter cells used for cms testing to come from a continous overall density distribution function (similar to cluster definitions).
filter.locmin is only applied, if k-min is specified as parameter in \code{\link{cms.cell}} or \code{\link{cms.pca}}.
}
\examples{

}
\seealso{
\code{\link{cms.cell}} for cms calculation

Other cms functions: \code{\link{cms.cell}},
  \code{\link{cms.integration}}, \code{\link{cms.pca}},
  \code{\link{cms.summary}}, \code{\link{smooth.cms}}
}
\concept{cms functions}
