% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cms.R
\name{smooth_cms}
\alias{smooth_cms}
\title{smooth_cms}
\usage{
smooth_cms(knn, cms_raw, cell_names, kmin, k)
}
\arguments{
\item{knn}{List containing distances and batch information of cells. Should have one slot type double named "distance" with distances towards the k nearest neighbours of the cell and one slot type double named after the group variable specified in 'batch'.
Both slots need to be aligned and the specified cell's name need to correspond to one of the rownames, while the columns represent the ordered knn.}

\item{cms_raw}{Matrix containing raw cms scores for all cells specified in cell_names and knn. Rownames need to correspond to cellnames and colnames should be "cms".}

\item{cell_names}{Character vector with cell names corresponding to the rownames of the list elements in the knn object and the cms_raw object.}

\item{kmin}{Numeric giving the minimum number of k nearest neighbours to use.
If k_min is defined the overall distance density distribution of knn will be determined and only cells up to the first local minima that includes more than kmin cells will be used.}

\item{k}{Numeric, defining the number of k-nearest neighbours to use for testing the "mixing" within the neighbourhood.}
}
\value{

}
\description{
Performs weighted smoothening of cms scores
}
\details{
Internal function to smoothes cms scores. In a complete random setting cms scores are uniform distributed.
To reduce the resulting random variance and enable visualization of local tendencies and pattern cms scores can be smoothened assuming that within one region mixing is uniform.
Generates smoothened cms scores using weigthed means of cms scores within the k-nearest neighbourhood.
Reciprocal distances are used as weights. Returns a matrix with "smoothened_cms" and original "cms" values.
}
\examples{

}
\seealso{
\code{\link{cms_cell}}, \code{\link{cms}}

Other cms functions: \code{\link{cms_cell}},
  \code{\link{cms_summary}}, \code{\link{cms}},
  \code{\link{filter_locmin}}
}
\concept{cms functions}
