% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{vis.cms}
\alias{vis.cms}
\title{vis.cms}
\usage{
vis.cms(cms_res, sce, cms_var = "cms", dim_red = "TSNE",
  log10_val = FALSE, ...)
}
\arguments{
\item{cms_res}{cms result matrix with one or two columns containing cms scores and/or smoothened cms scores.
Rownames need to correspond to colnames of sce. Colnames are allowed to be set different, but need to be specified in cms_var.}

\item{sce}{Combined \code{\link{SingleCellExperiment}} object containing either precalculated reduced dimension embeddings or counts or logcounts of all groups merged.
If precalculated dimension reduction embeddings are used, they need to be within the reducedDimensions slot.}

\item{cms_var}{character string specifying the cms scores to use (usually "cms_smoothened" or "cms").
Needs to correspond to one of the colnames of cms_res. Default is "cms".}

\item{dim_red}{Character defining dimension reduction embeddings to plot. Default is TSNE as from \code{\link{runTSNE}}.
If none is provided and no dimension reduction slot named "TSNE" is within \code{\link{reducedDimNames}} tsne embeddings will be calculated by \code{\link{runTSNE}}.}

\item{log10_val}{A logical indicating if -log10(cms) should be plotted to visualize differences of small values}

\item{...}{Additional arguments to pass to \code{\link{ggplot}}}
}
\description{
Plot cms scores in a reduced dimensional plot like tsne.
}
\details{
Plots a reduced dimension plot (tsne as default) colored by cms scores.
The dimesion reduction embedding can be specified, but only tsne embeddings will automatically be computed.
Embeddings from data integration methods (e.g. mnn.correct) can also been used as long as they are specified in \code{\link{reducedDimNames}} of sce.
}
\examples{

}
\seealso{
\code{\link{vis.overview}}, \code{\link{vis.group}}

Other visualize cms functions: \code{\link{compare.cluster}},
  \code{\link{compare.integration}},
  \code{\link{vis.group}}, \code{\link{vis.hist}},
  \code{\link{vis.overview}}
}
\concept{visualize cms functions}
