% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{vis.group}
\alias{vis.group}
\title{vis.group}
\usage{
vis.group(sce, group, dim_red = "TSNE", ...)
}
\arguments{
\item{sce}{Combined \code{\link{SingleCellExperiment}} object containing either precalculated reduced dimension embeddings or counts or logcounts of all groups merged.
If precalculated dimension reduction embeddings are used, they need to be within the reducedDimensions slot.}

\item{group}{character string specifying the group variable to label cells. Should be included into the colData slot of sce.}

\item{dim_red}{Character defining dimension reduction embeddings to plot. Default is TSNE as from \code{\link{runTSNE}}.
If none is provided and no dimension reduction slot named "TSNE" is within \code{\link{reducedDimNames}} tsne embeddings will be calculated by \code{\link{runTSNE}}.}

\item{...}{Additional arguments to pass to \code{\link{ggplot}}}
}
\description{
Plot group label in a reduced dimensional plot like tsne or integrated embeddings.
}
\details{
Plots a reduced dimension plot (tsne as default) colored by group parameter.
The dimesion reduction embedding can be specified, but only tsne embeddings will automatically be computed.
Embeddings from data integration methods (e.g. mnn.correct) can also been used as long as they are specified in \code{\link{reducedDimNames}} of sce.
}
\examples{

}
\seealso{
\code{\link{vis.overview}}, \code{\link{vis.cms}}

Other visualize cms functions: \code{\link{compare.cluster}},
  \code{\link{compare.integration}}, \code{\link{vis.cms}},
  \code{\link{vis.hist}}, \code{\link{vis.overview}}
}
\concept{visualize cms functions}
