% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{vis.hist}
\alias{vis.hist}
\title{vis.hist}
\usage{
vis.hist(cms, ...)
}
\arguments{
\item{cms}{cms result matrix with one or two columns containing cms scores and/or scaled cms scores.}

\item{...}{Additional arguments to pass to \code{\link{ggplot2}}.}
}
\description{
Plot pvalue histograms of cms score distributions
}
\details{
Plots cms score distribution similar to a pvalue histogram distribution.
Scaled cms should be close to a normal distribution centered around 0.5 and unscaled should be approx. flat, if no dataset-specific bias are expected.
}
\examples{

}
\seealso{
Other visualize cms functions: \code{\link{compare.cluster}},
  \code{\link{compare.integration}}, \code{\link{vis.cms}},
  \code{\link{vis.group}}, \code{\link{vis.overview}}
}
\concept{visualize cms functions}
