% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{vis.overview}
\alias{vis.overview}
\title{vis.overview}
\usage{
vis.overview(cms_res, sce, group, dim_red = "TSNE",
  scaled = ifelse(length(colnames(cms_res)) > 1, TRUE, FALSE),
  log10_val = FALSE, other_Var = NULL, ...)
}
\arguments{
\item{cms_res}{cms result matrix with one or two columns containing cms scores and/or scaled cms scores.
Rownames need to correspond to colnames of sce.}

\item{sce}{Combined \code{\link{SingleCellExperiment}} object containing either precalculated reduced dimension embeddings or counts or logcounts of all groups merged.
If precalculated dimension reduction embeddings are used, they need to be within the reducedDimensions slot.}

\item{group}{Character string specifying the name of variable used to define groups (batches). Should be have a corresonding element in colData(sce).}

\item{dim_red}{Character defining dimension reduction embeddings to plot. Default is TSNE as from \code{\link{runTSNE}}.
If none is provided and no dimension reduction slot named "TSNE" is within \code{\link{reducedDimNames}} tsne embeddings will be calculated by \code{\link{runTSNE}}.}

\item{scaled}{A logical value indicating if scaled cms scores should be plotted.
If true cms_res needs to contain 2 colums with cms and cms_scaled results. Default is set by the length of colnames(cms_res)}

\item{log10_val}{A logical indicating if -log10(cms) should be plotted to visualize differences of small values}

\item{other_Var}{Character string defining other variables to be plotted asided (e.g. some specified celltype label).
Need to be specified in colData(sce).}

\item{...}{Additional arguments to pass to \code{\link{runTSNE}} (CHECK!!!)}
}
\description{
Plot cms, scaled cms, group label and other Variable defined in colData in a reduced dimensional representation.
}
\details{
Plots reduced dimensions (tsne as default) of cells using the group variable and the cms score as colors.
Other color label as celltype label or scaled cms scores can be plotted aside. Generates tsne embeddings, if none have been specified.
Embeddings from data integration methods (e.g. mnn.correct) can also been used as long as they are specified in \code{\link{reducedDimNames}}.
}
\examples{

}
\seealso{
\code{\link{vis.cms}}, \code{\link{vis.group}}

Other visualize cms functions: \code{\link{compare.cluster}},
  \code{\link{compare.integration}}, \code{\link{vis.cms}},
  \code{\link{vis.group}}, \code{\link{vis.hist}}
}
\concept{visualize cms functions}
