% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisCms.R
\name{visCms}
\alias{visCms}
\title{visCms}
\usage{
visCms(cms_res, sce, cms_var = "cms", dim_red = "TSNE",
  log10_val = FALSE, ...)
}
\arguments{
\item{cms_res}{Matrix or data.frame. Cms scores to plot should be in columns and cells in rows.}

\item{sce}{A \code{SingleCellExperiment} object with the combined data corresponding to 'cms_res'.}

\item{cms_var}{Character Name of the cms scores to use (usually "cms_smoothened" or "cms").}

\item{dim_red}{Character. Name of embeddings to use as subspace for plotting. Default is "TSNE".}

\item{log10_val}{Logical. Indicating if -log10(cms) should be plotted.}

\item{...}{Additional arguments to pass to \code{\link{ggplot}}.}
}
\value{
a \code{ggplot} object.
}
\description{
Plot cms scores in a reduced dimensional plot.
}
\details{
Plots a reduced dimension plot colored by cms scores.
The dimesion reduction embedding can be specified, but only tsne embeddings will automatically be computed using \code{runTSNE}.
Embeddings from data integration methods (e.g. mnn.correct) can be used as long as they are specified in \code{reducedDimNames(sce)}.
}
\examples{
library(SingleCellExperiment)
load(system.file("extdata/sim30.rda", package = "CellMixS"))
load(system.file("extdata/cms_sim30.rda", package = "CellMixS"))
sce <- sim_30[[1]][, c(1:50, 500:550)]

visCms(cms_res = cms_sim30, sce = sce)

}
\seealso{
\code{\link{visOverview}}, \code{\link{visGroup}}

Other visualize cms functions: \code{\link{compareCluster}},
  \code{\link{compareIntegration}}, \code{\link{visGroup}},
  \code{\link{visHist}}, \code{\link{visOverview}}
}
\concept{visualize cms functions}
