% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisCms.R
\name{visHist}
\alias{visHist}
\title{visHist}
\usage{
visHist(cms_res)
}
\arguments{
\item{cms_res}{cms result matrix with one or two columns containing cms scores and/or smoothened cms scores.}
}
\description{
Plot pvalue histograms of cms score distributions
}
\details{
Plots cms score distribution similar to a pvalue histogram distribution.
Smoothened cms should be close to a normal distribution centered around 0.5 and cms scores should be approx. flat, if no dataset-specific bias are expected.
}
\examples{

}
\seealso{
Other visualize cms functions: \code{\link{compareCluster}},
  \code{\link{compareIntegration}}, \code{\link{visCms}},
  \code{\link{visGroup}}, \code{\link{visOverview}}
}
\concept{visualize cms functions}
