% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisCms.R
\name{visOverview}
\alias{visOverview}
\title{visOverview}
\usage{
visOverview(cms_res, sce, group, dim_red = "TSNE",
  smooth = ifelse(length(colnames(cms_res)) > 1, TRUE, FALSE),
  log10_val = FALSE, other_Var = NULL)
}
\arguments{
\item{cms_res}{Matrix or data.frame. Cms scores to plot should be in columns and cells in rows.}

\item{sce}{A \code{SingleCellExperiment} object with the combined data corresponding to 'cms_res'.}

\item{group}{Character. Name of group/batch variable. Needs to be one of \code{names(colData(sce))}.}

\item{dim_red}{Character. Name of embeddings to use as subspace for plotting. Default is "TSNE".}

\item{smooth}{Logical. Indicating if smoothened cms scores should be plotted.}

\item{log10_val}{Logical. Indicating if -log10(cms) should be plotted.}

\item{other_Var}{Character string. Name of other variables to be plotted asided.
Need correspond to one of \code{colData(sce)}.}
}
\value{
a \code{ggplot} object.
}
\description{
Plot an overview of cms, smoothened cms, group label and any colData variable in a reduced dimensional representation.
}
\details{
Plots reduced dimensions of cells colored by group variable and cms score.
If 'red_dim' is not defined in \code{reducedDimNames(sce)} a tsne is calculated using \code{runTSNE}.
Other color label as celltype label or smoothened cms scores can be plotted aside.
Embeddings from data integration methods (e.g. mnn.correct) can be used as long as they are specified in \code{reducedDimNames(sce)}.
}
\examples{
library(SingleCellExperiment)
load(system.file("extdata/sim30.rda", package = "CellMixS"))
load(system.file("extdata/cms_sim30.rda", package = "CellMixS"))
sce <- sim_30[[1]][, c(1:50, 500:550)]

visOverview(cms_sim30, sce, "batch")


}
\seealso{
\code{\link{visCms}}, \code{\link{visGroup}}

Other visualize cms functions: \code{\link{compareCluster}},
  \code{\link{compareIntegration}}, \code{\link{visCms}},
  \code{\link{visGroup}}, \code{\link{visHist}}
}
\concept{visualize cms functions}
