% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cms.R
\name{cms.pca}
\alias{cms.pca}
\title{cms.pca}
\usage{
cms.pca(sce, k, group, k_min = NA, smooth = TRUE, n_pc = 30,
  cell_min = 10)
}
\arguments{
\item{sce}{\code{\link{SingleCellExperiment}} object containing all cells of interest.}

\item{k}{Numeric, defining the number of k-nearest neighbours to use for testing the "mixing" within the neighbourhood.}

\item{group}{Character string specifying the name of variable used to define groups (batches). Should be have a corresonding element in knn.}

\item{k_min}{Numeric giving the minimum number of k nearest neighbours to use. cells to include.
If k_min is defined the overall distance density distribution of knn will be determined and only cells up to the first local minima that includes more than kmin cells will be used.}

\item{smooth}{A logical value indicating if cms results should be smoothened within each neighbourhood.}

\item{n_pc}{Numeric giving the number of PCs to include to define the subspace to calculate cell distances in.}

\item{cell_min}{Minimum number of cells from each group to be included into the AD test for calculation of cms.
Should be > 10 to make the ad.test function working.}
}
\description{
Calculates cell-specific mixing scores based on euclidean distances within a PC defined subspace for all cells of an SingleCellExperiment object
}
\details{
Specified version of \code{\link{cms.integration}}. The cms tests the hypothesis, that group-specific distance distributions of the k-nearest neighbouring cells have the same underlying unspecified distribution using the Anderson-Darling test as implemented in \code{\link{cms.cell}}.
In default the function uses all distances and group label defined in knn. If *k_min* is specified, the first local minimum of the overall distance distribution with at least kmin cells is used.
This can be used to adapt to the local structure of the datatset e.g. prevent cells from a distinct different cluster to be included, while keeping the overall number of k nearest neighbours reasonable large.
cms.pca uses a *SingleCellExperiment* \code{\link{SingleCellExperiment}} object as input and calculates euclidean distances within the subspace defined by the n_pc argument (default = 30).
For *smoothening* weigthed means of cms scores within each cell's k-nearest neighbourhood are calculated.
Reciprocal distances are used as weights.
}
\examples{

}
\seealso{
\code{\link{cms.cell}} for cms, \code{\link{smooth.cms}} for smoothing.

Other cms functions: \code{\link{cms.cell}},
  \code{\link{cms.integration}}, \code{\link{cms.summary}},
  \code{\link{filter.locmin}}, \code{\link{smooth.cms}}
}
\concept{cms functions}
