% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmsSummary.R
\name{cmsSummary}
\alias{cmsSummary}
\title{cmsSummary}
\usage{
cmsSummary(cms_res, sum_var = NULL, sce = NULL)
}
\arguments{
\item{cms_res}{data frame of cms scores to be summarized. Columns should correspond to cms scores, rows to cells.}

\item{sum_var}{variable to group summaries on. Need to correspond to a colData variable}

\item{sce}{SingleCellExperiment object describing the cells specified in cms_res.}
}
\description{
Summarizes cms scores, if specified based on groups
}
\details{
Returns the mean of cms score to make different conditions/methods/.. comparable.
In default the mean of each cloumn of cms_res is returned. Groups to summarize can be specified by sum_var and sce.
}
\examples{

}
\seealso{
\code{\link{compareIntegration}}, \code{\link{compareCluster}}

Other cms functions: \code{\link{cmsCell}},
  \code{\link{cms}}, \code{\link{filterLocMin}},
  \code{\link{smoothCms}}
}
\concept{cms functions}
