% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{compare.cluster}
\alias{compare.cluster}
\title{compare.cluster}
\usage{
compare.cluster(cms_res, cluster_var, cms_var = "cms", sce = NULL)
}
\arguments{
\item{cms_res}{data frame of cms scores to be summarized. The cms score should be in the first coulmn. It should either contain a second column with corresponding levels for cluster_var or  sce need to be specified.}

\item{cluster_var}{character string specifying the name of the factor level variable to summarize cms scores on.}

\item{sce}{SingleCellexperiment object should only be specified if cluster-var is not already provided in cms_res. Should contain 'cluster_var' within colData.
If sce is specified, rownames of cms_res need to correspond to colnames of sce (should contain the same cells).}
}
\description{
Creates summary boxplots of cms scores for different groups/cluster.
}
\details{
Plots summarized cms scores for a specified list. This function is intended to visualize and compare cms scores among clusters or other dataset variables.
}
\examples{

}
\seealso{
\code{\link{compare.integration}}

Other plot cms functions: \code{\link{compare.integration}},
  \code{\link{plot.cms}}, \code{\link{plot.group}},
  \code{\link{plot.hist}}, \code{\link{plot.overview}}
}
\concept{plot cms functions}
