% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmsSummary.R
\name{compareCluster}
\alias{compareCluster}
\title{compareCluster}
\usage{
compareCluster(cms_res, cluster_var, cms_var = "cms", sce = NULL,
  violin = FALSE)
}
\arguments{
\item{cms_res}{data frame of cms scores to be summarized. The cms score should be in the first coulmn. It should either contain a second column with corresponding levels for cluster_var or  sce need to be specified.}

\item{cluster_var}{character string specifying the name of the factor level variable to summarize cms scores on.}

\item{cms_var}{Character string specifying the name of the cms_res to use. Default is "cms".}

\item{sce}{SingleCellexperiment object should only be specified if cluster_var is not already provided in cms_res. Should contain 'cluster_var' within colData.
If sce is specified, rownames of cms_res need to correspond to colnames of sce (should contain the same cells).}

\item{violin}{A logical, if true (default) violin plots are provided. Otherwise boxplots are generated.}
}
\description{
Creates summary violin plots of cms scores for different groups/cluster.
}
\details{
Plots summarized cms scores for a specified list. This function is intended to visualize and compare cms scores among clusters or other dataset variables.
}
\examples{

}
\seealso{
\code{\link{compareIntegration}}

Other visualize cms functions: \code{\link{compareIntegration}},
  \code{\link{visCms}}, \code{\link{visGroup}},
  \code{\link{visHist}}, \code{\link{visOverview}}
}
\concept{visualize cms functions}
