% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisCms.R
\name{visHist}
\alias{visHist}
\title{visHist}
\usage{
visHist(cms_res, ncol = ifelse(length(colnames(cms_res)) > 1, 2, 1))
}
\arguments{
\item{cms_res}{Matrix or data.frame. Cms scores to plot should be in columns and cells in rows.}

\item{ncol}{Numeric. Number of columns of the pval histogram.}
}
\value{
a \code{ggplot} object.
}
\description{
Plot pvalue histograms of cms score distributions
}
\details{
Plots cms score distribution similar to a pvalue histogram distribution.
Without dataset-specific bias, cms scores should be approx. flat distributed.
}
\examples{
load(system.file("extdata/cms_sim30.rda", package = "CellMixS"))
visHist(cms_sim30)

}
\seealso{
Other visualize cms functions: \code{\link{compareCluster}},
  \code{\link{compareIntegration}}, \code{\link{visCms}},
  \code{\link{visGroup}}, \code{\link{visOverview}}
}
\concept{visualize cms functions}
